[12.0][FIX] l10n_it_delivery_note: fix migration for draft TDs
The module l10n_it_ddt creates TDs in draft with some fields that are
initially empty. (they are only marked as required in certain views)

When those fields are empty, the migration fails.
However, those fields are not required by l10n_it_delivery_note.

This commit simply adds a check to certain fields so that the migration
is performed correctly even when these fields are empty.